% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbuildr.R
\docType{package}
\name{ggbuildr}
\alias{ggbuildr}
\alias{ggbuildr-package}
\title{ggbuildr: A package for incrementally building layers of a ggplot}
\description{
The \code{ggbuildr} package provides functions for saving incremental
versions of a ggplot object, mainly to be used in a slide deck.
}
